osx build: don't set deployment target
If a deployment target is set within CMakeLists.txt detection might fail
if CMAKE_OSX_SYSROOT isn't set.
In case no deployment target is specified the latest available SDK is used as
deployment target and sysroot is set accordingly.

An other deployment target can still be set with -DCMAKE_OSX_DEPLOYMENT_TARGET=10.x
when running cmake for the first time or can be overwritten in cmake cache
later on.