Need Week instead of Day for TZ transistion
transistion.Day is being pulled and assigned to the DST transition times instead of the week of transition.  This gives the wrong date for DST implementation.
Pull request #1713 can be killed if libfreerdp/locale/timezone.c can be regenerated form the file resulting from the execution of this script.

All implementations of DST appear to be calculated rather than specific. 
http://www.webexhibits.org/daylightsaving/g.html
so transistion.Day should never be relevant.