Make remdesk_virtual_channel_write() static
The function with the same name is defined in multiple libraries
(libfreerdp-shadow.so, libfreerdp-server.so), which might confuse
the dynamic linker at runtime, binding a wrong version of the function
and thus causing segmentation faults.

Since remdesk_virtual_channel_write()s aren't used outside the files
they are defined in, we can declare them static to make them invisible
to the linker.